Added Laravel
I added the laravel folder. It's all set up except for the .env file,
which you have to maintain. Also, I put it in a seperate directory for
ease of use. If you don't want to hook things into laravel yourselves,
just let me know and put it in a directory outside of it. I can handle
it from there :D