Hooked up frontend, added question model
I've added the frontend views to Laravel. Note that I put the images
inside the public folder instead of storage for ease of use; this will
probably be the way we do it for this project just because our images
are static (usually).
I also added the question model. It includes a question string, a
suggestion string, and references to left child, right child, and
parent. I had to make these direct references, instead of one-to-many
references, because of the order structure imposed by a binary tree.
I'll go over how to use this model when we meet next.